app/backend/requirements.txt
fastapi
uvicorn
pydantic
google-genai
google-cloud-storage
requests
python-dotenv
httpx
pytest
feedparser
